requirements.txt
python-telegram-bot
httpx
beautifulsoup4
lxml